+++ src/hex_utils.c
@@ -18,7 +18,7 @@ uint8_t hex_parseNibblePair(const char* buffer)
 {
 	uint8_t first = hex_parseNibble(buffer[0]);
 	uint8_t second = hex_parseNibble(buffer[1]);
-	return (first << 4) + second;
+	return (uint8_t) ((first << 4) + second);
 }
 
 void stream_appendFromHexString(stream_t* s, const char* inStr)
